# [REVIEW] Optimize  `PG.add_data`